requirements.txt
mal-api.py
requests
